Refactor - create get teamInfo  mehtod & move BASE_URL api to constant
- https://github.com/next-step/js-todo-list-step3/pull/8#discussion_r481148539
- https://github.com/next-step/js-todo-list-step3/pull/8#discussion_r481152918
- Todo/App 컴포넌트에서 팀의 정보를 가져오는 api 호출에 대해 함수 처리
- BASE_URL이라는 값이 export하고 있지만 실제적으론 파일 내부에서 사용되는 문제가 있음
- 따라서, BASE_URL 자체를 constants.js로 옮겨 관리
