+++ src/js/utils/api/index.js
@@ -1,8 +1,6 @@
 import { default as teamAPI } from './team.js';
 import { default as userAPI } from './user.js';
 
-export const BASE_URL =
-  'https://blackcoffee-todolist.df.r.appspot.com/api/teams';
 export const request = async (url, option = {}) => {
   const response = await fetch(url, option);
   if (!response.ok) {
